% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_languages.R
\name{available_languages}
\alias{available_languages}
\title{List supported languages of DeepL API Pro}
\usage{
available_languages(auth_key = "your_key")
}
\arguments{
\item{auth_key}{authentication key.}
}
\description{
\code{available_languages} list all supported languages of DeepL API Pro.
}
\details{
To get an authentication key, you need to register for a DeepL API Pro
    account (\url{https://www.deepl.com/pro#developer}).
}
\examples{
\dontrun{
available_languages(auth_key = "my_key")
}

}
\references{
\href{https://www.deepl.com/pro#developer}{DeepL API documentations}
}
