% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_glossary_ff.R
\name{create_glossary2}
\alias{create_glossary2}
\title{Create a DeepL Glossary}
\usage{
create_glossary2(name, source_lang, target_lang, entries, auth_key, ...)
}
\arguments{
\item{name}{A character string for the name of the glossary.}

\item{source_lang}{A character string for the source language code.}

\item{target_lang}{A character string for the target language code.}

\item{entries}{A named list of entries where names are source terms and values are target terms.}

\item{auth_key}{A character string with the DeepL authentication key.}
}
\value{
A list containing information about the created glossary, including its ID.
}
\description{
This function creates a new glossary in DeepL using the provided entries in a named list using the free api.
}
\examples{
\dontrun{
glossary <- create_glossary2(
  name = "my_glossary",
  source_lang = "EN",
  target_lang = "ES",
  entries = list("Hello" = "Hola", "World" = "Mundo"),
  auth_key = "your_auth_free_key"
)
}
}
