% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeplr-package.R
\docType{package}
\name{deeplr-package}
\alias{deeplr}
\alias{deeplr-package}
\title{Fork of the \href{https://github.com/zumbov2/deeplr}{\code{deeplr} Package}}
\description{
An R wrapper for the DeepL Translator API
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.deepl.com/translator}
  \item Report bugs at \url{https://github.com/zumbov2/deeplr/issues}
}

}
\author{
\strong{Maintainer}: David Zumbach \email{david.zumbach@gfzb.ch}

Authors:
\itemize{
  \item Paul C. Bauer \email{mail@paulcbauer.eu}
}

}
\keyword{internal}
