% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeplr.R
\docType{package}
\name{deeplr}
\alias{deeplr}
\title{\code{deeplr} package}
\description{
An R wrapper for the DeepL Translator API
}
\details{
See the README on
\href{https://github.com/zumbov2/deeplr#readme}{GitHub}
}
