% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_glossary_ff.R
\name{delete_glossary}
\alias{delete_glossary}
\title{Delete a DeepL Glossary}
\usage{
delete_glossary(glossary_id, auth_key, ...)
}
\arguments{
\item{glossary_id}{A named list of entries where names are source terms and values are target terms.}

\item{auth_key}{A character string with the DeepL authentication key.}
}
\value{
A boolean.
}
\description{
This function deletes a glossary in DeepL using the provided id using the paid api.
}
\examples{
\dontrun{
glossary <- delete_glossary(
  glossary_id = "id_of_glossary_to_delete",
  auth_key = "your_auth_key"
)
}
}
