% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_glossaries_ff.R
\name{list_glossaries2}
\alias{list_glossaries2}
\title{List all DeepL Glossaries of an auth_key}
\usage{
list_glossaries2(auth_key, ...)
}
\arguments{
\item{auth_key}{A character string with the DeepL authentication key.}
}
\value{
A named list with all glossaries
}
\description{
This function returns the glossaries in DeepL hosted on provided auth_key using the free api.
}
\examples{
\dontrun{
glossaries <- list_glossaries2(
  auth_key = "your_auth_key"
)
}
}
