% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate2_df.R
\name{translate2_df}
\alias{translate2_df}
\title{Translate a character vector of a data frame element by element using DeepL}
\usage{
translate2_df(data, column_name, source_lang = NULL,
  target_lang = "EN", get_detect = FALSE, subdomain = "www2")
}
\arguments{
\item{data}{a data frame.}

\item{column_name}{vector to be translated. A single element must not exceed 5000 characters. Only UTF8-encoded plain text
is supported. Elements may contain multiple sentences.}

\item{source_lang}{vector containing the source languages of the texts to be translated (see below). If argument \code{is.null},
the API will try to detect the languages of the source. If input is of length 1, the same source language is applied to all elements.}

\item{target_lang}{vector containing the languages into which to translate. If input is of length 1, all elements are translated
    into the same language. Can be one of the following:
\itemize{
\item \code{EN} English
\item \code{DE} German
\item \code{FR} French
\item \code{ES} Spanish
\item \code{IT} Italian
\item \code{NL} Dutch
\item \code{PL} Polish
 }}

\item{get_detect}{if \code{TRUE}, the languages detected for the source texts are also included in the response. It corresponds to
the values of the argument \code{source_lang} if it was specified. If \code{FALSE}, only the translated texts are returned.}

\item{subdomain}{specifies the deepl subdomain to be used for the translation request. Currently: 'www2'}
}
\value{
If \code{get_detect} is set to \code{FALSE} a \code{character vector} containing the translations
    is returned. Otherwise, a \code{data.frame} (\code{tibble::tibble}) is returned with the following columns:
\itemize{
\item \code{translation} the translated texts.
\item \code{source_lang} detected or specified languages of the input texts.
}
}
\description{
\code{translate2_df} translates a character vector of a given data frame between English, German, French, Spanish, Italian,
    Dutch and Polish using the undocumented JSON-RPC DeepL API. No authentication key is required to use this service.
}
\examples{
\donttest{
# Data frame
dat <- data.frame(
  text = c("La seule facon de savoir ce qui se passe est de perturberle systeme.",
           "The whole problem with the world is that fools are always so certain of themselves"
           ), stringsAsFactors = F)

# Translate
translate2_df(dat, text, target_lang = "DE")
}

}
