% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_df.R
\name{translate_df}
\alias{translate_df}
\title{Translates a column in a dataframe.}
\usage{
translate_df(dataset = NULL, column.name = NULL, source.lang = "DE",
  target.lang = "EN", auth_key = NULL,
  url = "https://api.deepl.com/v1/translate?text=")
}
\arguments{
\item{dataset}{Dataframe with column of class character.}

\item{column.name}{Name of column that should be translated.}

\item{source.lang}{Source language. Either "detect" or one of "EN", "FR", etc.}

\item{target.lang}{Target language. One of "EN", "FR", etc.}

\item{auth_key}{Your API key.}

\item{url}{Url to DeepL API.}
}
\value{
Dataframe with original columns, translated column and column indicated detected or set source language.
}
\description{
Translates a column in a dataframe.
}
\examples{
\dontrun{
dat <- data.frame(text = c("La seule facon de savoir ce qui se passe est de perturber le systeme.",
                           "The whole problem with the world is that fools are always so certain of themselves"))

 translate_df(dataset = dat,
              column.name = "text",
              source.lang = "EN",
              target.lang = "DE",
              auth_key = "enter you auth key here")

dat <- data.frame(text = c("A dog.",
                           "Un chien.",
                           "Un perro.",
                           "Un cane.",
                           "Een hond."))

 translate_df(dataset = dat,
              column.name = "text",
              source.lang = "detect",
              target.lang = "DE",
              auth_key = "enter you auth key here")
}
}
