% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_languages2.R
\name{available_languages2}
\alias{available_languages2}
\title{List supported languages of DeepL API Free}
\usage{
available_languages2(auth_key = "your_key")
}
\arguments{
\item{auth_key}{authentication key.}
}
\description{
\code{available_languages2} list all supported languages of DeepL API Free.
}
\details{
To get an authentication key, you need to register for a DeepL API Free
    account (\url{https://www.deepl.com/pro#developer}).
}
\examples{
\dontrun{
available_languages2(auth_key = "my_key")
}

}
\references{
\href{https://www.deepl.com/pro#developer}{DeepL API documentations}
}
