% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeplr-package.R, R/deeplr.R
\docType{package}
\name{deeplr-package}
\alias{deeplr-package}
\title{deeplr: Interface to the 'DeepL' Translation API}
\description{
A wrapper for the 'DeepL' Pro API \url{https://www.deepl.com/docs-api}, a web service for translating texts between different languages. A DeepL API developer account is required to use the service (see \url{https://www.deepl.com/pro#developer}).

An R wrapper for the DeepL Translator API
}
\details{
See the README on
\href{https://github.com/zumbov2/deeplr#readme}{GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.deepl.com/translator}
  \item Report bugs at \url{https://github.com/zumbov2/deeplr/issues}
}


Useful links:
\itemize{
  \item \url{https://www.deepl.com/translator}
  \item Report bugs at \url{https://github.com/zumbov2/deeplr/issues}
}

}
\author{
\strong{Maintainer}: David Zumbach \email{david.zumbach@gfzb.ch}

Authors:
\itemize{
  \item Paul C. Bauer \email{mail@paulcbauer.eu}
}

}
\keyword{internal}
