% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeplr.R
\docType{package}
\name{deeplr-package}
\alias{deeplr}
\alias{deeplr-package}
\title{\code{deeplr} package}
\description{
An R wrapper for the DeepL Translator API
}
\details{
See the README on
\href{https://github.com/zumbov2/deeplr#readme}{GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.deepl.com/translator}
  \item Report bugs at \url{https://github.com/zumbov2/deeplr/issues}
}

}
\author{
\strong{Maintainer}: David Zumbach \email{david.zumbach@gfzb.ch}

Authors:
\itemize{
  \item Paul C. Bauer \email{mail@paulcbauer.eu}
}

}
