% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detect}
\alias{detect}
\title{Detect the language of a text using the official DeepL Translator API}
\usage{
detect(text, auth_key = "your_key")
}
\arguments{
\item{text}{text to be translated. Only UTF8-encoded plain text is supported. May contain multiple sentences.
The request size should not exceed 30kbytes.}

\item{auth_key}{DeepL authentication key which provides access to the API.}
}
\description{
\code{detect} detects the language of a text using the official DeepL Translator API. English, German, French,
    Spanish, Italian, Dutch and Polish are currently available. To use this service, an authentication key is required.
}
\examples{
\donttest{
detect("My name is Hans.", auth_key = "my_key")

}
}
