% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect2.R
\name{detect2}
\alias{detect2}
\title{Detect the language of a text using DeepL}
\usage{
detect2(text, subdomain = "www2")
}
\arguments{
\item{text}{text to be translated. Must not exceed 5000 characters. Only UTF8-encoded plain text is supported.
May contain multiple sentences.}

\item{subdomain}{specifies the deepl subdomain to be used for the translation request. Currently: 'www2'}
}
\description{
\code{detect2} detects the language of a text using the undocumented JSON-RPC DeepL API. English, German, French,
    Spanish, Italian, Dutch and Polish are currently available. No authentication key is required to use this service.
}
\examples{
\donttest{
detect2("My name is Hans.")

}
}
