% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pimp2.R
\name{pimp2}
\alias{pimp2}
\title{Fix and improve texts using DeepL}
\usage{
pimp2(text, help_lang)
}
\arguments{
\item{text}{text to be translated. Must not exceed 5000 characters. Only UTF8-encoded plain text is supported.
May contain multiple sentences.}

\item{help_lang}{language used as a help language for reverse translation. Can be one of the following:
\itemize{
\item \code{EN} English
\item \code{DE} German
\item \code{FR} French
\item \code{ES} Spanish
\item \code{IT} Italian
\item \code{NL} Dutch
\item \code{PL} Polish
 }}
}
\description{
\code{pimp2} translates a text into a help language and then back into the original language
    using the undocumented JSON-RPC DeepL API. No authentication key is required to use this service.
}
\examples{
\donttest{
pimp2("In former times I lived in Zurich", help_lang = "DE")
}

}
