% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_df.R
\name{translate_df}
\alias{translate_df}
\title{Translates a column in a dataframe.}
\usage{
translate_df(dataset = NULL, column.name = NULL, source.lang = "DE",
  target.lang = "EN", add.source.lang = FALSE, auth_key = NULL,
  url = "https://api.deepl.com/v1/translate?text=")
}
\arguments{
\item{dataset}{Dataframe with column of class character.}

\item{column.name}{Name of column that should be translated.}

\item{source.lang}{Source language. Either "detect" or one of "EN", "FR", etc.}

\item{target.lang}{Target language. One of "EN", "FR", etc.}

\item{add.source.lang}{A logical. If TRUE adds the preset or detected language for each
translation. FALSE (default) adds nothing.}

\item{auth_key}{Your API key.}

\item{url}{Url to DeepL API.}
}
\value{
Dataframe with original columns and translated column. add.source.lang = TRUE adds
additional column "source_lang".
}
\description{
Translates a column in a dataframe.
}
\examples{
\dontrun{
  dat <- data.frame(text = c("La seule facon de savoir ce qui se passe est de perturber
                              le systeme.",
                              "The whole problem with the world is that fools are always so
                              certain of themselves"))


# Without language detection and without adding set/detected source language.
  translate_df(dataset = dat,
               column.name = "text",
               source.lang = "EN",
               target.lang = "DE",
               auth_key = "enter you auth key here")

  dat <- data.frame(text = c("A dog.",
                             "Un chien.",
                             "Un perro.",
                             "Un cane.",
                             "Een hond."))

# With language detection and with adding set/detected source language.
  translate_df(dataset = dat,
               column.name = "text",
               source.lang = "detect",
               target.lang = "DE",
               add.source.lang = TRUE,
               auth_key = "enter you auth key here")
}
}
