% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_vec.R
\name{translate_vec}
\alias{translate_vec}
\title{Translates a character vector.}
\usage{
translate_vec(dataset = NULL, source.lang = "DE", target.lang = "EN",
  auth_key = NULL, url = "https://api.deepl.com/v1/translate?text=")
}
\arguments{
\item{dataset}{A character vector.}

\item{source.lang}{Source language. Either "detect" or one of "EN", "FR", etc.}

\item{target.lang}{Target language. One of "EN", "FR", etc.}

\item{auth_key}{Your API key.}

\item{url}{Url to DeepL API.}
}
\value{
A character vector. If source.lang = "detect" also the detected languages.
}
\description{
Translates a character vector.
}
\examples{
\dontrun{
dat <- c("La seule facon de savoir ce qui se passe est de perturber le systeme.",
         "The whole problem with the world is that fools are always so certain of themselves"))

 translate_vec(dataset = dat,
              source.lang = "EN",
              target.lang = "DE",
              auth_key = "enter you auth key here")

dat <- c("A dog.",
         "Un chien.",
         "Un perro.",
         "Un cane.",
         "Een hond.")

 translate_vec(dataset = dat,
              source.lang = "detect",
              target.lang = "DE",
              auth_key = "enter you auth key here")
}
}
