% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage_ff.R
\name{usage}
\alias{usage}
\title{Usage data of a DeepL API Pro account}
\usage{
usage(auth_key = "your_key")
}
\arguments{
\item{auth_key}{authentication key of the corresponding DeepL API Pro account.}
}
\description{
\code{usage} returns the character usage and the configured limit for the current period of
    a DeepL API Pro account.
}
\details{
To get an authentication key, you need to register for a DeepL API Pro
    account (\url{https://www.deepl.com/pro#developer}).
}
\examples{
\dontrun{
usage(auth_key = "my_key")
}

}
\references{
\href{https://www.deepl.com/pro#developer}{DeepL API documentations}
}
