% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage.R
\name{usage}
\alias{usage}
\title{Retrieve current usage data of a DeepL Pro account}
\usage{
usage(auth_key = "your_key")
}
\arguments{
\item{auth_key}{authentication key of the corresponding DeepL Pro account.}
}
\description{
\code{usage} returns the character usage and the configured limit for the current period of a DeepL Pro Account.
}
\details{
To get an authentication key, you need to register for a DeepL Pro account (\url{https://www.deepl.com/pro.html}).
    This currently costs 20 euros per month and allows the translation of 1,000,000 characters per month (see
    \url{https://www.deepl.com/pro-pricing.html}).
}
\examples{
\donttest{
# Simple translation
usage(auth_key = "my_key")

}


}
\references{
\href{https://www.deepl.com/api.html}{DeepL API documentations}
}
