% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage2.R
\name{usage2}
\alias{usage2}
\title{Usage data of a DeepL API Free account}
\usage{
usage2(auth_key = "your_key")
}
\arguments{
\item{auth_key}{authentication key.}
}
\description{
\code{usage2} returns the character usage and the configured limit for the current period of
    a DeepL API Free account.
}
\details{
To get an authentication key, you need to register for a DeepL API Pro
    account (\url{https://www.deepl.com/pro#developer}).
}
\examples{
\dontrun{
usage(auth_key = "my_key")
}

}
\references{
\href{https://www.deepl.com/pro#developer}{DeepL API documentations}
}
